+++ src/app/BaseLayout/Header.tsx
@@ -2,12 +2,10 @@ import "server-only"
 
 import { FC } from "react"
 import { AppShellHeader } from "./AppShell"
-import { Flex, Indicator, Space } from "@mantine/core"
+import { Flex } from "@mantine/core"
 import { getSession } from "@/auth/server/auth"
 import HeaderDrawerMenu from "./HeaderDrawerMenu"
 import { HeaderAvatar } from "./HeaderAvatar"
-import { ActionIcon } from "@/components/ActionIcon"
-import { NotificationIcon } from "@/components/icon/Notification"
 import { LoginButton } from "@/components/LoginButton"
 import Image from "next/image"
 import LogoImage from "@/../public/logo-rect.png"
@@ -38,21 +36,9 @@ const Header: FC<HeaderProps> = async () => {
                 <Flex align="center">
                     {session
                         ? <>
-                            <Indicator>
-                                <ActionIcon
-                                    size="lg"
-                                    variant="subtle"
-                                    radius="xl"
-                                    color="info"
-
-                                >
-                                    <NotificationIcon />
-                                </ActionIcon>
-                                <Space w="0.5em" />
-                                <HeaderAvatar
-                                    session={session}
-                                />
-                            </Indicator>
+                            <HeaderAvatar
+                                session={session}
+                            />
                         </>
                         : <LoginButton />
                     }
